+++ mercury/system/po.py
@@ -28,8 +28,6 @@ class PostOffice:
     Convenient class for making RPC, async and callback.
     """
 
-    DEFERRED_DELIVERY = 'system.deferred.delivery'
-
     def __init__(self):
         self.platform = Platform()
         self.util = Utility()
@@ -68,20 +66,29 @@ class PostOffice:
             event.set_body(body)
         self.platform.send_event(event, True)
 
-    def send_later(self, route: str, headers: dict = None, body: any = None, seconds: float = 1.0) -> None:
+    def send_later(self, route: str, headers: dict = None, body: any = None, reply_to: str = None, me=True,
+                   seconds: float = 1.0) -> None:
         self.util.validate_service_name(route, True)
         if isinstance(seconds, float) or isinstance(seconds, int):
-            relay = dict()
-            relay['route'] = route
-            if headers is not None:
-                relay_headers = dict()
-                for h in headers:
-                    relay_headers[str(h)] = str(headers[h])
-                relay['headers'] = relay_headers
-            if body is not None:
-                relay['body'] = body
-            relay['seconds'] = seconds
-            self.send(self.DEFERRED_DELIVERY, body=relay)
+            if seconds > 0:
+                if headers is None and body is None:
+                    raise ValueError('Unable to send because both headers and body are missing')
+                event = EventEnvelope().set_to(route)
+                if headers is not None:
+                    if not isinstance(headers, dict):
+                        raise ValueError('headers must be dict')
+                    for h in headers:
+                        event.set_header(str(h), str(headers[h]))
+                if body is not None:
+                    event.set_body(body)
+                if reply_to is not None:
+                    if not isinstance(reply_to, str):
+                        raise ValueError('reply_to must be str')
+                    # encode 'me' in the "call back" if replying to this instance
+                    event.set_reply_to(reply_to, me)
+                self.platform.send_event_later(seconds, event)
+            else:
+                raise ValueError('delay in seconds must be larger than zero')
         else:
             raise ValueError('delay in seconds must be int or float')