+++ CHANGELOG.md
@@ -6,11 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 2.3.4, 5/18/2022
+
+### Added
+
+Support deferred delivery in standalone mode
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 2.3.3, 5/6/2022
 
 ### Added
 
-for unhandled exception, return exception class name and error message.
+For unhandled exception, return exception class name and error message.
 
 ### Removed